freeze connection policies and connection mappings
The current way of doing things is way too dangerous, and quite
a few times already has caused objects that were shared to be
modified, without it being desired.

Freeze all connection mappings and connection policies to ensure
that nothing stupid is being done. This requires to fix the way
the dataflow dynamics update the connection policies, which is
really for the best.