cleanup the deployment management API
This commit modifies the way deployments are managed, from a global
API on the Configuration singleton, to a more flexible API based
on "deployment group" objects.

Deployments can now be specified / selected locally. Existing
disambiguation methods are retained for backward compatibility, but
should be replaced in new code by the new deployment API.

Each deployment group defines a set of deployments, that can then
be selected at the InstanceRequirements level. Groups on a given task
(say, selected on a composition) apply on the whole subnet (say, the
composition's tasks) unless there is another explicitely selected
group in the hierarchy

This commit does not contain the profile-level API