models: move DeploymentGroup conf and loaders argument to the relevant methods
Storing them makes droby-dumping problematic as conf/loaders are
really not meant to be dumped. Given that they are resolution
arguments, not part of the groups' state, and that I really felt
they were a small, wanted to get rid of them in the first place,
that's a step in the right direction.

In fine, we probably should get rid of them completely (requiring
the caller to have already loaded the relevant models).