refactor PlaceholderTask
This commit normalizes how placeholder are calls within the syskit
code base, that is placeholders and not proxies (they are definitely
not proxying anything).

PlaceholderTask is split between Models::Placeholder and Placeholder,
the same way than the rest of the API. In addition, we make sure to
use the #placeholder? predicate to check whether a model or instance
are a placeholder or not.

The weird split of the creation API between Syskit singleton methods
and methods at the component level is also fixed to have everything
under Syskit::Models::Placeholder. A new method
(Placeholder.new_specialized_placeholder) allows to create
specialized placeholders (such as profile tags) in a clean way.

And fix some bugs found thanks to the newly written unit test suite (some
potentially pretty bad)